create styles folder in src directory, successfully extracted login Form styles into that folder
