+++ src/styles/login/loginFormStyles.js
@@ -1,0 +1,30 @@
+import React, { Component } from 'react';
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    padding: 20
+  },
+  input: {
+    height: 40,
+    backgroundColor: 'rgba(255,255,255,0.2)',
+    marginBottom: 15,
+    color: '#FFF',
+    paddingHorizontal: 10
+  },
+  buttonLogin: {
+    backgroundColor: '#9b59b6',
+    paddingVertical: 8,
+    marginBottom: 30
+  },
+  buttonCreate: {
+    backgroundColor: '#8e44ad',
+    paddingVertical: 8
+  },
+  buttonText: {
+    textAlign: 'center',
+    color: '#FFF'
+  }
+})
+
+export default styles;
